phone is set instead of userId

+++ zhem-nails-service/src/main/java/ru/zhem/repository/UserRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import ru.zhem.entity.User;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, BigDecimal> {
 }
